refactor(rdf-core): the xsd_regex translator folds over the scanner instead of its own cursor
`translate`, `translate_escape` and `translate_unicode_property` move to a new
private `emit` module and are rewritten as a flat fold over the `Scanner` token
stream introduced in the previous commit: no `chars.get(i)`, no manual index
advance, no lookahead. All cursor state now lives in `scan.rs`, so the duplicate
`ClassFrame` and `Groups` copies are deleted from `translate.rs`, which drops
from 801 to 110 lines and now holds only `ecma_262_divergences`.

The negated-class subtraction binding is preserved without rebuilding the class
stack in the emitter. The scanner keeps owning `ClassFrame` and exposes a
per-token `closes_negated_wrap()` hint, so the emitter knows whether `]` owes
one closing bracket or two and whether `-[` must close the inner negated class
before emitting `--`. `[^0-9-[a-z]]` still emits `[[^0-9]--[a-z]]`, `[^abc]`
still emits `[[^abc]]`, and nested subtraction is unchanged.

Behaviour is byte-identical. A temporary differential test compared old and new
translation over 616 distinct inputs (all six .cases corpus files, 118
hand-written patterns covering negated classes, nested subtraction, block
escapes, \i/\c, groups, backreferences, dot and the q/x flags, plus every
pattern re-run through strip_x_flag_whitespace) across both dot_all settings:
1232 comparisons with zero divergence in emitted source, error variant, or
error text. The temporary test was deleted before commit.
